feat: improve button design
