+++ client/src/components/TextToSpeech/TextToSpeech.tsx
@@ -1,85 +1,63 @@
+// src/components/TextToSpeech/TextToSpeech.tsx
 import { useState } from 'react';
-import { AVAILABLE_VOICES } from '../../config/voices';
 import { generateSpeech } from '../../services/text-to-speech';
-import { useAudio } from '../../hooks/useAudio';
-import { VoiceSelector } from './VoiceSelector';
+import { findMatchingVoice } from '../../utils/voiceUtils';
+import textToSpeechIcon from '../../assets/img/text-to-speech.svg';
+import './main.css';
 
-export const TextToSpeech = () => {
-  const [text, setText] = useState('');
-  const [isGenerating, setIsGenerating] = useState(false);
-  const [selectedVoice, setSelectedVoice] = useState(AVAILABLE_VOICES[0].id);
-  const [error, setError] = useState<string | null>(null);
-  
-  const { audioElement, isPlaying, play, togglePlayPause } = useAudio();
+interface TextToSpeechProps {
+  beastName: string;
+  text: string;
+}
 
-  const handleGenerateAudio = async () => {
-    if (!text.trim()) return;
+export const TextToSpeech = ({ beastName, text }: TextToSpeechProps) => {
+  const [isPlaying, setIsPlaying] = useState(false);
+  const [currentAudio, setCurrentAudio] = useState<HTMLAudioElement | null>(null);
 
-    setIsGenerating(true);
-    setError(null);
+  const handleStop = () => {
+    if (currentAudio) {
+      currentAudio.pause();
+      setCurrentAudio(null);
+      setIsPlaying(false);
+    }
+  };
 
+  const handlePlay = async () => {
     try {
-      const data = await generateSpeech(text, selectedVoice);
+      if (currentAudio) {
+        currentAudio.pause();
+        setIsPlaying(false);
+      }
+
+      const beastVoice = findMatchingVoice(beastName);
+      const data = await generateSpeech(text, beastVoice.id);
+      
       if (data.audio) {
-        play(data.audio);
+        const audio = new Audio(`data:audio/mpeg;base64,${data.audio}`);
+        setCurrentAudio(audio);
+        
+        audio.onended = () => {
+          setIsPlaying(false);
+          setCurrentAudio(null);
+        };
+
+        audio.play();
+        setIsPlaying(true);
       }
     } catch (error) {
-      if (error instanceof Error) {
-        setError(error.message);
-      }
-    } finally {
-      setIsGenerating(false);
+      console.error('Error playing audio:', error);
     }
   };
 
   return (
-    <div className="bg-white p-6 rounded-lg shadow-md">
-      <div className="space-y-4">
-        <VoiceSelector
-          selectedVoice={selectedVoice}
-          voices={AVAILABLE_VOICES}
-          onVoiceChange={setSelectedVoice}
-        />
-
-        <div>
-          <label className="block text-sm font-medium mb-2">Enter Text</label>
-          <textarea
-            value={text}
-            onChange={(e) => setText(e.target.value)}
-            placeholder="Enter the text you want to convert to speech..."
-            className="w-full h-32 p-2 border rounded-md resize-none"
-          />
-        </div>
-
-        {error && (
-          <div className="text-red-500 text-sm p-2 bg-red-50 rounded-md">
-            {error}
-          </div>
-        )}
-
-        <div className="flex gap-2">
-          <button
-            onClick={handleGenerateAudio}
-            disabled={isGenerating || !text.trim()}
-            className={`flex-1 px-4 py-2 rounded-md text-white ${
-              isGenerating || !text.trim()
-                ? "bg-gray-400"
-                : "bg-blue-500 hover:bg-blue-600"
-            }`}
-          >
-            {isGenerating ? "Generating..." : "Generate Audio"}
-          </button>
-
-          {audioElement && (
-            <button
-              onClick={togglePlayPause}
-              className="px-4 py-2 border rounded-md hover:bg-gray-100"
-            >
-              {isPlaying ? "Pause" : "Play"}
-            </button>
-          )}
-        </div>
+    <button
+      className="audio-button"
+      onClick={() => isPlaying ? handleStop() : handlePlay()}
+      title={`Listen to ${beastName}'s description`}
+    >
+      <div>
+        <img src={textToSpeechIcon} alt="text to speech" />
       </div>
-    </div>
+    </button>
   );
 };
